Change naming conventions, bug fixes.
As discussed on claw-dev, changed many of the names of Makefile
variables.  Also fixed a problem with multiple library paths and
including them in the ALL_INCLUDE flags.

Note that in order for this change to work you must have the variable
CLAW_DIR set to the Clawpack base path.  This was also discussed on
claw-dev and is open to comment.