Add preprocessor directives
There now exists a set of preprocessor pragmas in this commits files
that turn on and off the OpenMP pragmas embedded in the files.  The
Makefile for the AMR swirl example has been modified to include
additional compiler flags to run the preprocessor in the case of
gfortran and to trigger the grid level threading that previously
existed.