Changed max1d to 600, added comment about OpenMP
The max1d parameter controls the maximum size of each grid.  This has
been set to 600 in the past but with the use of OpenMP grid level
looping, it was set to 60 for load balancing.  If you intend to use the
OpenMP grid level looping, reset this parameter to something around 60
for better load balancing.